+++ modules/migration/hooks/useMigrationStatus.tsx
@@ -2,18 +2,12 @@ import { useDelegatedTo } from 'modules/delegates/hooks/useDelegatedTo';
 import { useActiveWeb3React } from 'modules/web3/hooks/useActiveWeb3React';
 import { useDelegateContractExpirationDate } from 'modules/delegates/hooks/useDelegateContractExpirationDate';
 import { isAboutToExpireCheck, isExpiredCheck } from '../helpers/expirationChecks';
-import { addressConnections } from 'modules/migration/connections';
 
 export function useMigrationStatus(): {
   isDelegatedToExpiredContract: boolean;
   isDelegatedToExpiringContract: boolean;
   isDelegateContractExpired: boolean;
   isDelegateContractExpiring: boolean;
-  newOwnerAddress?: string;
-  newOwnerConnected: boolean;
-  previousOwnerAddress?: string;
-  previousOwnerConnected: boolean;
-  newOwnerHasDelegateContract: boolean;
 } {
   const { account: address, network } = useActiveWeb3React();
 
@@ -40,31 +34,10 @@ export function useMigrationStatus(): {
       }, false)
     : false;
 
-  const oldToNewMap = addressConnections;
-  const newToOldMap = Object.keys(addressConnections).reduce((acc, cur) => {
-    return {
-      ...acc,
-      [addressConnections[cur]]: cur
-    };
-  }, {});
-
-  const previousOwnerConnected = address ? !!oldToNewMap[address] : false;
-  const newOwnerConnected = address ? !!newToOldMap[address] : false;
-
-  const previousOwnerAddress = previousOwnerConnected ? address : address ? newToOldMap[address] : undefined;
-  const newOwnerAddress = newOwnerConnected ? address : address ? oldToNewMap[address] : undefined;
-
-  const newOwnerHasDelegateContract = !!delegateContractExpirationDate;
-
   return {
     isDelegatedToExpiredContract,
     isDelegatedToExpiringContract,
     isDelegateContractExpired,
-    isDelegateContractExpiring: true,
-    newOwnerAddress,
-    newOwnerConnected,
-    previousOwnerAddress,
-    previousOwnerConnected,
-    newOwnerHasDelegateContract
+    isDelegateContractExpiring: true
   };
 }
